requirements.txt
absl_py==0.15.0
biopython==1.80
dm_tree==0.1.7
fair_esm==2.0.0
ml_collections==0.1.1
numpy==1.24.2
scipy>=1.10.0
torch==1.13.1
pandas==1.5.3
torch_cluster==1.6.1+pt113cu116
torch_geometric==2.3.1
torch_scatter==2.1.1+pt113cu116
tqdm==4.64.1